refactor: Mindmap_tab 관련 타입 분리
